requirements.txt
torch
torchvision
numpy
matplotlib
scikit-image
Flask
Pillow 